Componentize the TablatureSizeSlider. This is the functionality that allows the user to adjust the size of the tablature
